# Fix per-run combine time (fixes #181)