--- a/Criterion/Measurement.hs
+++ b/Criterion/Measurement.hs
@@ -196,6 +196,8 @@ measure bm iters = runBenchmarkable bm iters addResults $ \ !n act -> do
            }
   return (m, endTime)
   where
+    -- When combining runs, the Measured value is accumulated over many runs,
+    -- but the Double value is the most recent absolute measurement of time.
     addResults :: (Measured, Double) -> (Measured, Double) -> (Measured, Double)
     addResults (!m1, _) (!m2, !d2) = (m3, d2)
       where